requirements.txt
alembic==1.2.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dash==1.3.1
dash-core-components==1.2.1
dash-html-components==1.0.1
dash-renderer==1.1.0
dash-table==4.3.0
decorator==4.4.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-core==4.5.0
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.17.2
pandas==0.25.1
pandas-datareader==0.8.0
plotly==4.1.1
psycopg2==2.8.4
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
retrying==1.3.3
six==1.12.0
SQLAlchemy==1.3.8
traitlets==4.3.2
urllib3==1.25.5
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
python-dotenv==0.13.0
